refactor(webkit): move legacy components into wip/ (PR 5/7)
Quarantines 11 pre-spec components under src/components/wip/ so the new
spec-driven layers (actions/, content/, inputs/, …) have a clean root.
Public API (the @aziontech/webkit/<name> subpaths in package.json) is
unchanged — only the source file location moves.

Folders moved (each is a pure git rename, no content edits):
  azion-system-status, box-grid-selection, buttons (6 button-*),
  empty-results-block, overline, pricing-card, resizable-splitter,
  segmented-button, site (site/button), tag, testimonials-carousel

Touched:
- packages/webkit/package.json: 16 export paths repointed at wip/
- packages/webkit/src/core/list-data-table/cells/cell-{expand,clipboard}.vue:
  CopyBlock import now uses wip/buttons/button-copy
- packages/webkit/src/core/list-data-table/components/list-data-table/list-data-table.vue:
  EmptyResultsBlock import now uses wip/empty-results-block
- packages/webkit/src/core/card/card-content/card-content.vue:
  Button import now uses wip/site/button

Smoke tests
- pnpm webkit:lint passes
- pnpm --filter storybook build completes; tokens (--bg-canvas, --primary,
  .gap-spacing-elements-md, .display-1) still present in preview-*.css
