+++ packages/webkit/package.json
@@ -76,23 +76,23 @@
     "./field-text-privacy": "./src/core/form/field-text-privacy/field-text-privacy.vue",
     "./label": "./src/core/form/label/label.vue",
     "./selector-block": "./src/core/selector-block/selector-block.vue",
-    "./azion-system-status": "./src/components/azion-system-status/azion-system-status.vue",
-    "./testimonials-carousel": "./src/components/testimonials-carousel/testimonials-carousel.vue",
-    "./empty-results-block": "./src/components/empty-results-block/empty-results-block.vue",
-    "./resizable-splitter": "./src/components/resizable-splitter/resizable-splitter.vue",
-    "./overline": "./src/components/overline/overline.vue",
-    "./box-grid-selection": "./src/components/box-grid-selection/box-grid-selection.vue",
-    "./pricing-card": "./src/components/pricing-card/pricing-card.vue",
+    "./azion-system-status": "./src/components/wip/azion-system-status/azion-system-status.vue",
+    "./testimonials-carousel": "./src/components/wip/testimonials-carousel/testimonials-carousel.vue",
+    "./empty-results-block": "./src/components/wip/empty-results-block/empty-results-block.vue",
+    "./resizable-splitter": "./src/components/wip/resizable-splitter/resizable-splitter.vue",
+    "./overline": "./src/components/wip/overline/overline.vue",
+    "./box-grid-selection": "./src/components/wip/box-grid-selection/box-grid-selection.vue",
+    "./pricing-card": "./src/components/wip/pricing-card/pricing-card.vue",
     "./accordion": "./src/core/primevue/accordion/accordion.vue",
     "./accordion-tab": "./src/core/primevue/accordion-tab/accordion-tab.vue",
     "./avatar": "./src/core/primevue/avatar/avatar.vue",
     "./button": "./src/core/primevue/button/button.vue",
-    "./button-cancel": "./src/components/buttons/button-cancel/button-cancel.vue",
-    "./button-clone": "./src/components/buttons/button-clone/button-clone.vue",
-    "./button-copy": "./src/components/buttons/button-copy/button-copy.vue",
-    "./button-create": "./src/components/buttons/button-create/button-create.vue",
-    "./button-delete": "./src/components/buttons/button-delete/button-delete.vue",
-    "./button-save": "./src/components/buttons/button-save/button-save.vue",
+    "./button-cancel": "./src/components/wip/buttons/button-cancel/button-cancel.vue",
+    "./button-clone": "./src/components/wip/buttons/button-clone/button-clone.vue",
+    "./button-copy": "./src/components/wip/buttons/button-copy/button-copy.vue",
+    "./button-create": "./src/components/wip/buttons/button-create/button-create.vue",
+    "./button-delete": "./src/components/wip/buttons/button-delete/button-delete.vue",
+    "./button-save": "./src/components/wip/buttons/button-save/button-save.vue",
     "./breadcrumb": "./src/core/primevue/breadcrumb/breadcrumb.vue",
     "./card": "./src/core/primevue/card/card.vue",
     "./card-box": "./src/core/card/card-box/card-box.vue",
@@ -117,7 +117,7 @@
     "./tabmenu": "./src/core/primevue/tabmenu/tabmenu.vue",
     "./tabpanel": "./src/core/primevue/tabpanel/tabpanel.vue",
     "./tabview": "./src/core/primevue/tabview/tabview.vue",
-    "./tag": "./src/components/tag/tag.vue",
+    "./tag": "./src/components/wip/tag/tag.vue",
     "./prime-tag": "./src/core/primevue/tag/tag.vue",
     "./timeline": "./src/core/primevue/timeline/timeline.vue",
     "./toast": "./src/core/primevue/toast/toast.vue",
@@ -137,8 +137,8 @@
     "./listbox": "./src/core/primevue/listbox/listbox.vue",
     "./calendar": "./src/core/primevue/calendar/calendar.vue",
     "./picklist": "./src/core/primevue/picklist/picklist.vue",
-    "./segmented-button": "./src/components/segmented-button/segmented-button.vue",
-    "./site/button": "./src/components/site/button/button.vue",
+    "./segmented-button": "./src/components/wip/segmented-button/segmented-button.vue",
+    "./site/button": "./src/components/wip/site/button/button.vue",
     "./actions/button": "./src/components/actions/button/button.vue",
     "./actions/icon-button": "./src/components/actions/icon-button/icon-button.vue",
     "./utils/spinner": "./src/components/utils/spinner/spinner.vue",
