+++ packages/webkit/src/components/wip/resizable-splitter/package.json
