+++ packages/webkit/src/core/card/card-content/card-content.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
   import { computed } from 'vue'
 
-  import Button from '../../../components/site/button/button.vue'
+  import Button from '../../../components/wip/site/button/button.vue'
 
   defineOptions({ name: 'CardContent' })
